+++ src/@tests/functional/auth/session-lifecycle.test.ts
@@ -33,7 +33,7 @@ Deno.test({
   fn: async () => {
     Deno.env.set('REDIS_URI', 'redis://localhost:6379')
     Deno.env.set('JWT_KEY', 'my-secret')
-    await import('jsr:@zanix/datamaster@0.5.*/core') // load cache core
+    await import('@zanix/datamaster/core') // load cache core
     await ProgramModule.connectors.get<any>('cache:redis').clear() // reset data
 
     const cache = ProgramModule.providers.get<ZanixCacheProvider>('cache')